--- uade_battleship/menu-improved.py
@@ -1,172 +1,0 @@
-import os
-import pygame, sys
-
-from uade_battleship.board import board
-
-pygame.init()
-
-SCREEN = pygame.display.set_mode((1280, 720))
-pygame.display.set_caption("Menu")
-
-BG = pygame.image.load("assets/Background.png")
-
-
-def get_font(size):
-    return pygame.font.Font("assets/font.ttf", size)
-
-
-def play():
-    board()
-
-
-def options():
-    while True:
-        OPTIONS_MOUSE_POS = pygame.mouse.get_pos()
-
-        SCREEN.fill("white")
-
-        OPTIONS_TEXT = get_font(30).render("This is the OPTIONS screen.", True, "Black")
-        OPTIONS_RECT = OPTIONS_TEXT.get_rect(center=(640, 260))
-        SCREEN.blit(OPTIONS_TEXT, OPTIONS_RECT)
-
-        OPTIONS_BACK = Button(
-            image=None,
-            pos=(640, 460),
-            text_input="BACK",
-            font=get_font(30),
-            base_color="Black",
-            hovering_color="Green",
-        )
-
-        OPTIONS_BACK.changeColor(OPTIONS_MOUSE_POS)
-        OPTIONS_BACK.update(SCREEN)
-
-        for event in pygame.event.get():
-            if event.type == pygame.QUIT:
-                pygame.quit()
-                sys.exit()
-            if event.type == pygame.MOUSEBUTTONDOWN:
-                if OPTIONS_BACK.checkForInput(OPTIONS_MOUSE_POS):
-                    main_menu()
-
-        pygame.display.update()
-
-
-def main_menu():
-    while True:
-        SCREEN.blit(BG, (0, 0))
-
-        MENU_MOUSE_POS = pygame.mouse.get_pos()
-
-        MENU_TEXT = get_font(60).render("MENU", True, "#b68f40")
-        MENU_RECT = MENU_TEXT.get_rect(center=(640, 100))
-
-        PLAY_BUTTON = Button(
-            image=None,
-            pos=(640, 200),
-            text_input="Comenzar partida",
-            font=get_font(30),
-            base_color="#d7fcd4",
-            hovering_color="White",
-        )
-        OPTIONS_BUTTON = Button(
-            image=None,
-            pos=(640, 300),
-            text_input="Instrucciones de juego",
-            font=get_font(30),
-            base_color="#d7fcd4",
-            hovering_color="White",
-        )
-        CREDITS_BUTTON = Button(
-            image=None,
-            pos=(640, 400),
-            text_input="Configuraciones",
-            font=get_font(30),
-            base_color="#d7fcd4",
-            hovering_color="White",
-        )
-        HELP_BUTTON = Button(
-            image=None,
-            pos=(640, 500),
-            text_input="Scores",
-            font=get_font(30),
-            base_color="#d7fcd4",
-            hovering_color="White",
-        )
-        QUIT_BUTTON = Button(
-            image=None,
-            pos=(640, 600),
-            text_input="Salir",
-            font=get_font(30),
-            base_color="#d7fcd4",
-            hovering_color="White",
-        )
-
-        SCREEN.blit(MENU_TEXT, MENU_RECT)
-
-        for button in [
-            PLAY_BUTTON,
-            OPTIONS_BUTTON,
-            CREDITS_BUTTON,
-            HELP_BUTTON,
-            QUIT_BUTTON,
-        ]:
-            button.changeColor(MENU_MOUSE_POS)
-            button.update(SCREEN)
-
-        for event in pygame.event.get():
-            if event.type == pygame.QUIT:
-                pygame.quit()
-                sys.exit()
-            if event.type == pygame.MOUSEBUTTONDOWN:
-                if PLAY_BUTTON.checkForInput(MENU_MOUSE_POS):
-                    play()
-                if OPTIONS_BUTTON.checkForInput(MENU_MOUSE_POS):
-                    options()
-                if QUIT_BUTTON.checkForInput(MENU_MOUSE_POS):
-                    pygame.quit()
-                    sys.exit()
-
-        pygame.display.update()
-
-
-# button
-class Button:
-    def __init__(self, image, pos, text_input, font, base_color, hovering_color):
-        self.image = image
-        self.x_pos = pos[0]
-        self.y_pos = pos[1]
-        self.font = font
-        self.base_color, self.hovering_color = base_color, hovering_color
-        self.text_input = text_input
-        self.text = self.font.render(self.text_input, True, self.base_color)
-        text_width, text_height = self.text.get_size()
-        # Adjust button size to fit the text
-        self.image = pygame.Surface((text_width + 20, text_height + 20))
-        self.image.fill((0, 0, 0, 0))  # Transparent background
-        self.rect = self.image.get_rect(center=(self.x_pos, self.y_pos))
-        self.text_rect = self.text.get_rect(center=(self.x_pos, self.y_pos))
-
-    def update(self, screen):
-        if self.image is not None:
-            screen.blit(self.image, self.rect)
-        screen.blit(self.text, self.text_rect)
-
-    def checkForInput(self, position):
-        if position[0] in range(self.rect.left, self.rect.right) and position[
-            1
-        ] in range(self.rect.top, self.rect.bottom):
-            return True
-        return False
-
-    def changeColor(self, position):
-        if position[0] in range(self.rect.left, self.rect.right) and position[
-            1
-        ] in range(self.rect.top, self.rect.bottom):
-            self.text = self.font.render(self.text_input, True, self.hovering_color)
-        else:
-            self.text = self.font.render(self.text_input, True, self.base_color)
-
-
-if __name__ == "__main__":
-    main_menu()
